Audio Added

+++ 2D Movement Template/2D Player Movement template/Assets/Scripts/PlayerMove.cs
@@ -14,7 +14,7 @@ public class PlayerMove : MonoBehaviour
     private int spriteVersion = 0;
     public Sprite[] sprites;
     private SpriteRenderer sr;
-    public int animSpeed;
+    public float animSpeed;
 
     
     void Start()
